Work on moor othello board

+++ Moor Othello/Moor Othello/OthelloMove.cpp
@@ -10,6 +10,6 @@ bool OthelloMove::operator==(const OthelloMove & rhs) {
 // TODO check this is the correct format
 OthelloMove::operator std::string() const {
 	std::ostringstream moveStr;
-	moveStr << "(" << mPosition.getRow() << "," << mPosition.getColumn() << ")";
+	moveStr << "(" << (int) mPosition.getRow() << "," << (int) mPosition.getColumn() << ")";
 	return moveStr.str();
 }
